Move environments into a separate module
The "meaning" module has been used as a kitchensink location for all
stuff during initial development. It's time to split it up. Move the
environment handling code into a separate module. Add some docs too.
