+++ src/libeval/lib.rs
@@ -15,3 +15,4 @@ extern crate reader;
 pub mod expression;
 pub mod expanders;
 pub mod meaning;
+pub mod environment;
