+++ src/libeval/meaning.rs
@@ -19,86 +19,7 @@ use reader::datum::{ScannedDatum, DatumValue};
 use reader::intern_pool::{Atom};
 
 use expression::{Expression, ExpressionKind, Literal, Variable, Arguments};
-
-pub struct Environment {
-    kind: EnvironmentKind,
-    variables: Vec<Variable>,
-    parent: Option<Rc<Environment>>,
-}
-
-#[derive(Eq, PartialEq)]
-enum EnvironmentKind {
-    Local,
-    Global,
-    Imported,
-}
-
-pub enum VariableKind {
-    Local {
-        depth: usize,
-        index: usize,
-    },
-    Global {
-        index: usize,
-    },
-    Imported {
-        index: usize,
-    },
-    Unresolved,
-}
-
-impl Environment {
-    pub fn new_local(variables: &[Variable], parent: &Rc<Environment>) -> Rc<Environment> {
-        Rc::new(Environment {
-            kind: EnvironmentKind::Local,
-            variables: variables.to_vec(),
-            parent: Some(parent.clone()),
-        })
-    }
-
-    pub fn new_global(variables: &[Variable], parent: &Rc<Environment>) -> Rc<Environment> {
-        assert!(parent.kind == EnvironmentKind::Imported);
-        Rc::new(Environment {
-            kind: EnvironmentKind::Global,
-            variables: variables.to_vec(),
-            parent: Some(parent.clone()),
-        })
-    }
-
-    pub fn new_imported(variables: &[Variable]) -> Rc<Environment> {
-        Rc::new(Environment {
-            kind: EnvironmentKind::Imported,
-            variables: variables.to_vec(),
-            parent: None,
-        })
-    }
-
-    pub fn resolve_variable(&self, name: Atom) -> VariableKind {
-        // First, try to resolve the name locally.
-        for (index, local) in self.variables.iter().enumerate() {
-            if name == local.name {
-                return match self.kind {
-                    EnvironmentKind::Local => VariableKind::Local { index, depth: 0 },
-                    EnvironmentKind::Global => VariableKind::Global { index },
-                    EnvironmentKind::Imported => VariableKind::Imported { index },
-                };
-            }
-        }
-
-        // If that fails then look into parent environment (if it's available).
-        if let Some(ref parent) = self.parent {
-            let mut variable = parent.resolve_variable(name);
-            // Bump the nesting depth for local variables.
-            if let VariableKind::Local { ref mut depth, .. } = variable {
-                *depth += 1;
-            }
-            return variable;
-        }
-
-        // The variable cannot be resolved if it is absent in all environments.
-        return VariableKind::Unresolved;
-    }
-}
+use environment::{Environment, VariableKind};
 
 pub struct Meaning {
     pub kind: MeaningKind,
